SSLVSCANER/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==22.3
apscheduler==3.10.4
lxml==4.9.3
python-dotenv==1.0.0
sqlmodel==0.0.14
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing dependencies
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.0

# Development dependencies
black==24.10.0
isort==5.13.2
flake8==7.1.1

# Optional dependencies for better performance
httpx==0.28.1
urllib3==2.5.0